renamed rpath to dashboard from /welcome to /dashboard; added SearchResults view and redirect upon user search
