--- client/src/components/FetchedPlaces/index.tsx
@@ -1,27 +1,0 @@
-import React from 'react'
-
-
-// const FetchedPlaces = () => {
-//     return (
-//         {fetchedPlaces.map((place: SearchData) => (
-//             <ul key={place.yelp_id}>
-//               <h3>{place.name}</h3>
-//               <li>
-//                 {place.address1} / {place.city}, {place.state}
-//               </li>
-//               <li>{place.phone}</li>
-//               <li>{place.rating}</li>
-//               <button
-//                 type="button"
-//                 onClick={(): void => {
-//                   savePlace(place);
-//                 }}
-//               >
-//                 Save place
-//               </button>
-//             </ul>
-//           ))}
-//     )
-// }
-
-// export default FetchedPlaces
