+++ client/src/pages/Login/index.tsx
@@ -1,12 +1,12 @@
-import React, { useState, useContext } from "react";
-import UserContext from "../../context/UserContext";
-import { UserCredentialsState } from "../../interfaces";
-import axios from "axios";
+import React, { useState, useContext } from 'react';
+import UserContext from '../../context/UserContext';
+import { UserCredentialsState } from '../../interfaces';
+import axios from 'axios';
 
 const Login: React.FC = (): JSX.Element => {
   const [credentials, setCredentials] = useState<UserCredentialsState>({
-    email: "",
-    password: "",
+    email: '',
+    password: '',
   });
 
   const { handleLogin } = useContext(UserContext);
@@ -14,13 +14,13 @@ const Login: React.FC = (): JSX.Element => {
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>): void => {
     e.preventDefault();
     if (!credentials.email || !credentials.password) {
-      return window.alert('You must enter your email and password to login.')
+      return window.alert('You must enter your email and password to login.');
     } else loginUser(credentials);
   };
 
   const loginUser = async (loginInfo: UserCredentialsState): Promise<void> => {
     try {
-      const response = await axios.post("/login", loginInfo);
+      const response = await axios.post('/login', loginInfo);
       if (response.data.error === true) throw new Error();
       if (handleLogin) {
         handleLogin(response.data.body.sessionToken);
@@ -31,27 +31,27 @@ const Login: React.FC = (): JSX.Element => {
   };
 
   return (
-    <div className="loginContainer">
+    <div className='loginContainer'>
       <form onSubmit={handleSubmit}>
-        <label htmlFor="email">Email: </label>
+        <label htmlFor='email'>Email: </label>
         <input
-          type="text"
-          name="email"
+          type='text'
+          name='email'
           value={credentials.email}
           onChange={(e: React.ChangeEvent<HTMLInputElement>): void =>
             setCredentials({ ...credentials, email: e.target.value })
           }
         />
-        <label htmlFor="password"> Password: </label>
+        <label htmlFor='password'> Password: </label>
         <input
-          type="text"
-          name="password"
+          type='text'
+          name='password'
           value={credentials.password}
           onChange={(e: React.ChangeEvent<HTMLInputElement>): void =>
             setCredentials({ ...credentials, password: e.target.value })
           }
         />
-        <button type="submit">Submit</button>
+        <button type='submit'>Submit</button>
       </form>
     </div>
   );
