+++ client/src/pages/NoMatch/index.tsx
@@ -8,7 +8,7 @@ const NoMatch: React.FC = (): JSX.Element => {
       <h3 className='noMatchText'>
         I'm sorry, the page you are looking for does not exist.
       </h3>
-      <Link className='backButton' to='/welcome'>
+      <Link className='backButton' to='/dashboard'>
         <button>Go Back</button>
       </Link>
     </div>
